==> Notebooks/II. Polarity and Emotion Recognition/1. Preparing Lexicons.ipynb <==
## I. Reading NRC Lexicon and Its Translations to Different Languages:

import pandas as pd
lexicon = pd.read_excel('NRCLexicon/NRC-Emotion-Lexicon-v0.92/NRC-Emotion-Lexicon-v0.92-InManyLanguages-web.xlsx')
lexicon.head()

lexicon.columns

## II. Preparing Lexicon Dictionaries:

### 1. Choosing Language:

language = 'French' # REPLACE LANGUAGE HERE

### 2. Reading corresponding Emotion/Sentiment vectors for words translated to chosen language:

language_lexicon = lexicon[[language+' Translation (Google Translate)','Positive', 'Negative','Anger','Anticipation', 'Disgust', 'Fear', 'Joy', 'Sadness', 'Surprise','Trust']]
language_lexicon.head()

### 3. Converting the words and their vectors into a dictionary:

d = {}
for i in range(0,len(language_lexicon)):
    d[language_lexicon.iloc[i][language+' Translation (Google Translate)']] = [language_lexicon.iloc[i]['Anger'],language_lexicon.iloc[i]['Anticipation'], language_lexicon.iloc[i]['Disgust'], language_lexicon.iloc[i]['Fear'], language_lexicon.iloc[i]['Joy'],language_lexicon.iloc[i]['Negative'],language_lexicon.iloc[i]['Positive'], language_lexicon.iloc[i]['Sadness'], language_lexicon.iloc[i]['Surprise'],language_lexicon.iloc[i]['Trust']]

d = {k: v for k, v in d.items() if v !=[0,0,0,0,0,0,0,0,0,0]}

### 4. Storing the dictionary in a pickle file for later reuse:

import pickle

with open('NRCLexicon/'+language+'/'+language.lower()+'_lexicon.pickle', 'wb') as handle:
    pickle.dump(d, handle, protocol=2)

with open('NRCLexicon/'+language+'/'+language.lower()+'_lexicon.pickle', 'rb') as handle:
    b = pickle.load(handle)
len(b)

## III. Preparing Sets of Representative Words:

import pandas as pd
lexicon = pd.read_csv('NRCLexicon/NRC-Emotion-Lexicon-v0.92/NRC-emotion-lexicon.txt',engine='python', names=['col'])
lexicon['Word'] = lexicon['col'].str.split().apply(lambda x: x[0])
lexicon['EmotionCategory'] = lexicon['col'].str.split().apply(lambda x: x[1])
lexicon['Score'] = lexicon['col'].str.split().apply(lambda x: x[-1])
del(lexicon['EmotionCategory'])
del(lexicon['col'])

from tqdm import tqdm
distinct_words = list(set(lexicon['Word']))
word_vectors = []
for i in tqdm(range(0,len(distinct_words))):
    vector = list(lexicon[lexicon['Word']==distinct_words[i]]['Score'])
    word_vectors.append((distinct_words[i],vector))

vector_dict = dict(word_vectors)
vector_dict['happy']
import numpy as np
#np.save('NRCLexicon/lexicon_dictionary.npy', vector_dict)
if 'aims' in vector_dict.keys():
    print ("yes")
list_ = [0 for i in range(0,10)]
list_
vector_dict['clap']

lexicon_dict = np.load('NRCLexicon/lexicon_dictionary.npy').item()
lexicon_dict['clap']

### 2. Computing Lists of Representative Words for each emotion/sentiment category:

representative_set = []
for i in range(0,10):
    representative_set_sub = []
    for word in d.keys():
        if d[word][i] == 1: 
            representative_set_sub.append(word)
    representative_set.append(representative_set_sub)
representative_set[0][0:15]

### 3. Storing the list in a dataframe where each column represents the set of representative words for a particular emotion/sentiment category:

lexicon_df = pd.DataFrame()
lexicon_df[0] = representative_set[0]
for i in range(1,10):
    df = pd.DataFrame()
    df[i] = representative_set[i]
    lexicon_df= pd.concat([lexicon_df,df],ignore_index=True,axis=1)
lexicon_df.columns = [['Anger','Anticipation','Disgust','Fear','Joy','Negative','Positive','Sadness','Surprise','Trust']]

lexicon_df.to_csv('NRCLexicon/'+language+'/'+language.lower()+'_nrc.csv',index=False)
lexicon_df.head(15)
