==> Notebooks/II. Polarity and Emotion Recognition/2. Polarity Simple Approach.ipynb <==
#### Simple Approach - Semtiment Analysis
In this notebook, we apply a simple approach to analyse the sentiments of the tweets. we use a textblob library which has a support for 3 languages, English, French, and German.. the top three frequent languages in Switzerland.

We do a simpe cleaning befor using TextBlob Sentiment Analyzer, such as removing https links, hashtags and mentions.

import pandas as pd
from textblob import TextBlob
import numpy as np
import  csv
import matplotlib.pyplot as plt
import matplotlib
%matplotlib inline 
import re

tweet=pd.read_csv('tweets_with_canton_language_short.csv')

len(tweet)

## text cleaning

def clean_up_text(text):
    '''
    remove https links, hashtags and mentions
    '''
    hashtag = re.compile(r'[#]\w*')
    https = re.compile(r'https?:\/\/[a-zA-z0-9\/#%\.]+')
    mention = re.compile(r'[@]\w*')
    text = re.sub(hashtag, '', text)
    text = re.sub(https, '', text)
    text = re.sub(mention, '', text)
    return text


# English

tweet_en=tweet[tweet['language']=='en']

len(tweet_en)

text_list=tweet_en.text.tolist()

claned_text=[]
for t in text_list:
    claned_text.append(clean_up_text(t))
    

def get_sentiment_polarity(text):
    blob = TextBlob(text)
    return blob.polarity, blob.subjectivity


def study_sentiment(polarity_list, mid_point=0):
    '''
    The polarity score is a float within the range [-1.0, 1.0].
    The subjectivity is a float within the range [0.0, 1.0] 
    where 0.0 is very objective and 1.0 is very subjective.
    '''
    polarity_array = np.array(polarity_list)
    mean = np.mean(polarity_array)
    std = np.std(polarity_array)
    frac_positive = sum(polarity_array>mid_point)*1./len(polarity_array)
    frac_negative = sum(polarity_array<mid_point)*1./len(polarity_array)
    return mean, std, frac_positive,frac_negative

sentiment_analysis = map(get_sentiment_polarity, claned_text)
polarity_list = map(lambda x:x[0], sentiment_analysis)
sentiment_list = map(lambda x:x[1], sentiment_analysis)

po=[]
i=0
for a in polarity_list:
    #print(a)
    po.append(a)
    i+=1
    if (i%100000 == 0):
        print(i)

len(po)

len(tweet_en)

tweet_en['polarity']=po

# French

tweet_fr=tweet[tweet['language']=='fr']

len(tweet_fr)

from textblob import Blobber
from textblob_fr import PatternTagger, PatternAnalyzer
tb = Blobber(pos_tagger=PatternTagger(), analyzer=PatternAnalyzer())
blob1 = tb(u"Quelle belle matinée")
blob1.sentiment




text_list=tweet_fr.text.tolist()
claned_text=[]
for t in text_list:
    claned_text.append(clean_up_text(t))
    

score=[]
i=0
for text in claned_text:  
    blob2 = tb(text)
    score.append(blob2.sentiment[0])
    i+=1
    if (i%100000 == 0):
        print(i)


tweet_fr['polarity']=score

# German

tweet_de=tweet[tweet['language']=='de']

len(tweet_de)

text_list=tweet_de.text.tolist()
claned_text=[]
for t in text_list:
    claned_text.append(clean_up_text(t))
    

from textblob_de import TextBlobDE as TextBlob

score=[]
i=0
for text in text_list:  
    blob = TextBlob(text)
    score.append(blob.sentiment[0])
    i+=1
    if (i%10000 == 0):
        print(i)


len(score)

tweet_de['polarity']=score

tweet_fr_en_de_polarity=pd.concat([tweet_en,tweet_fr,tweet_de])

tweet_fr_en_de_polarity=tweet_fr_en_de_polarity.reset_index()

len(tweet_fr_en_de_polarity)

tweet_fr_en_de_polarity.to_csv('tweet_fr_en_de_polarity.csv', index=None)
