==> Notebooks/II. Polarity and Emotion Recognition/3. Emotion Recognition Simple Approach.ipynb <==
#### Simple Approach for Emotion Analysis

In this notebook, we apply a simple approach to capture the emotions expressed in tweets.
first, we clean the tweet by removing https links,hashtags, and mentions. then we remove stop words.
after that we use the NRC Emotion lexicon to look for emotion mapping for each word in a tweet and generate an emotion vector with 8 entries for each tweet, each entry in the vector represent a single emotion.

import pandas as pd
from textblob import TextBlob
import numpy as np
import  csv
import matplotlib.pyplot as plt
import matplotlib
%matplotlib inline 
import re
from nltk.corpus import stopwords
import nltk

## text cleaning
def clean_up_text(text):
    '''
    remove https links, hashtags and mentions
    '''
    hashtag = re.compile(r'[#]\w*')
    https = re.compile(r'https?:\/\/[a-zA-z0-9\/#%\.]+')
    mention = re.compile(r'[@]\w*')
    text = re.sub(hashtag, '', text)
    text = re.sub(https, '', text)
    text = re.sub(mention, '', text)
    return text

def remove_stop_words(text_list,stop_word_language):
    print("remove stop words start")
    c_t=[]
    for text in text_list:
        t=clean_up_text(text)
        c_t.append(t)
        
    nostopwords_lower_list=[]
    i=0
    for text in c_t:
        words_list=nltk.tokenize.word_tokenize(text)
        filtered_words = [word for word in words_list if word not in stopwords.words(stop_word_language)]
        filtered_words=' '.join(filtered_words)
        nostopwords_lower_list.append(filtered_words.lower())
        if (i%10000 == 0):
            print(i)
        i+=1
    return nostopwords_lower_list

def emotion_analysis(tweet_file_path, language, stop_word_language):
    tweet=pd.read_csv(tweet_file_path)
    print("number of weets: ",len(tweet))    
    nostopwords_lower_list=tweet['cleaned_text'].tolist()
    lec=pd.read_excel('NRC-Emotion-Lexicon-v0.92-InManyLanguages-web.xlsx')
    lec_language=lec[[language,'Positive','Negative','Anger','Anticipation','Disgust','Fear','Joy','Sadness','Surprise','Trust']]
    lec_language[language]=lec_language[language].str.lower()
    lec_language=lec_language.drop_duplicates(language)
    lec_language=lec_language.set_index([language])
    print("start emotion analysis")
    ee=0
    emotion_final_list=[]
    for text in nostopwords_lower_list:
        s=str(text).split(' ')
        emotion_list=[]
        for ss in s:
            if ss in lec_language.index:
                row=lec_language.loc[ss]
                ll=[row[0],row[1],row[2],row[3],
                    row[4],row[5],
                    row[6],row[7],row[8],row[9]]
                emotion_list.append(ll)
        k=[sum(i) for i in zip(*emotion_list)]
        if(k==[]):
            emotion_final_list.append([0,0,0,0,0,0,0,0,0,0])
        else:
            emotion_final_list.append(k)
        ee+=1
        if (ee%100000 == 0):
            print(ee)
    
    tweet=pd.read_csv(tweet_file_path)
    tweet['cleaned_text']=nostopwords_lower_list
    
    headers=['Positive','Negative','Anger','Anticipation','Disgust','Fear','Joy','Sadness','Surprise','Trust']
    df = pd.DataFrame(emotion_final_list, columns=headers)

    tweet=pd.concat([tweet,df], axis=1)
    
    return tweet

# English Tweets

tweets_with_emptions=emotion_analysis('intermediate_data/en_cleaned.csv', 'English Word','english')

# French Tweets

tweets_with_emptions=emotion_analysis('intermediate_data/fr_cleaned.csv', 'French Translation (Google Translate)','french')

tweets_with_emptions.head()

tweets_with_emptions.to_csv('emotions_data/fr_final_emotion.csv',index=None)

# German Tweets

tweets_with_emotions=emotion_analysis('intermediate_data/de_cleaned.csv', 'German Translation (Google Translate)', 'german')

tweets_with_emotions.to_csv('emotions_data/de_final_emotion.csv',index=None)

# Italian Tweets

tweets_with_emotions=emotion_analysis('intermediate_data/it_cleaned.csv', 'Italian Translation (Google Translate)', 'italian')

tweets_with_emotions.to_csv('emotions_data/it_final_emotion.csv',index=None)

# Spanish Tweets

tweets_with_emptions=emotion_analysis('intermediate_data/es_cleaned.csv', 'Spanish Translation (Google Translate)', 'spanish')

tweets_with_emptions.to_csv('emotions_data/es_final_emotion.csv',index=None)

# portuguese Tweets

tweets_with_emptions=emotion_analysis('intermediate_data/pt_cleaned.csv', 'Portuguese Translation (Google Translate)', 'portuguese')

tweets_with_emptions.to_csv('emotions_data/pt_final_emotion.csv',index=None)

# Turkish Tweets

tweets_with_emptions=emotion_analysis('intermediate_data/tr_cleaned.csv', 'Turkish Translation (Google Translate)', 'turkish')

tweets_with_emptions.to_csv('emotions_data/tr_final_emotion.csv',index=None)

# Dutch Tweets

tweets_with_emptions=emotion_analysis('intermediate_data/nl_cleaned.csv', 'Dutch Translation (Google Translate)', 'dutch')

tweets_with_emptions.to_csv('emotions_data/nl_final_emotion.csv',index=None)

# Arabic Tweets

tweets_with_emptions=emotion_analysis('intermediate_data/ar_cleaned.csv', 'Arabic Translation (Google Translate)', 'arabic')

tweets_with_emptions.to_csv('emotions_data/ar_final_emotion.csv',index=None)

### Concatenate all emotion files:

import glob, os
os.chdir("emotions_data")


emotion_big_df=pd.DataFrame()
i=0
for file in glob.glob("*.csv"):
    name=file.split(".")[0]
    print(name)
    df=pd.read_csv(file)
    print(len(df))
    emotion_big_df=pd.concat([emotion_big_df,df], axis=0)
emotion_big_df.head()

len(emotion_big_df)

emotion_big_df.to_csv('all_tweets_with_emotions.csv', index=None)
