==> Notebooks/II. Polarity and Emotion Recognition/4. Rule Based Affective Representation of Tweets.ipynb <==
import pandas as pd
from EmotionAnalysis.DataSchemaExtractionParsing import *
from EmotionAnalysis.DataPreProcessing import *
from EmotionAnalysis.SentSemanticModule import *
from EmotionAnalysis.SentTweetModule import *
from EmotionAnalysis.SentSyntacticModule import *
import math

## Purpose of this notebook:
In this notebook, we demonstrate the different steps followed in order to come up with a refined representation of each tweet by following two principles:
* Word Qualification: application of stop word removal, part of speech tagging and named entity recognition and term normalization to keep good refined emotional candidates. 
* Inter-Word Relationships: Application of syntactic analysis to study three kinds of dependencies:
    * Negation Dependency: e.g. I am not happy 
    * Adjectival Dependency: 
    * Adverbial Dependency
    
#### NB:
This notebook makes direct calls to functions defined in EmotionAnalysis folder. Please refer to that in order to see details of implementation of different steps of the pipeline:
* Pre processing
* Syntactic Module
* Semantic Word Level Module
* Semantic Tweet Level Module

We have runned the same code on the whole dataset chunk by chunk with several variations depending on the language and the libraries available for that specific language. But, for demonstration purposes, in this notebook, we show the process for a small subset.

## I. Loading English Data 

english_tweets = pd.read_csv("../../Data/Sample Data/en_sample.csv",encoding = "ISO-8859-1",nrows=8000)
english_tweets.head()

Example of a tweet before applying any processing (we will use this example and several others to show how our steps refine an intermediary representation before applying any emotion recognition methodology):

english_tweets['text'].iloc[160]

english_tweets['text'][9]

## II. Preliminary Pre-processing: 

### 1. Replacing Special Categories:
We first start by dealing with some string patterns that are particular to the case of Twitter Data:
* Urls: we detect and remove data as they don't carry any emotional importance
* Digits: we remove them for the same reason
* Detecting @ instances with <username> and removing it
* Removing hashtag # sign and keeping the word after hashtag

replaced_categories = handle_special_categories(english_tweets)

replaced_categories['text'].iloc[160]

We notice here that url, username and hashtag sign have been removed

### 2. Replacing contractions (needed for more accurate tokenization)
e.g, "they're stunning" becomes "they are stunning". Otherwise, if we use tokenization right away, it will split into two words: they and 're and 're will cannot be removed since it is not part of stopwords.  

tweets_no_contractions = replace_contractions(replaced_categories)

tweets_no_contractions['text'][9]

Notice here "how we're" has been replaced by "how we are".

### 3. Tokenization of Tweets into words:

tokenized_list = bag_of_word_representation(tweets_no_contractions)
tokenized_list[10]

## III. Syntactic Analysis:

### 1. Part of Speech Tagging:
We use part of speech tagging here in order to detect N.A.V.A. words (Nouns, Adjectives, Verbs, Adverbs) those are good candidates to carry emotions.

tagged_tweets = pos_tagging(tokenized_list)
tagged_tweets[10]

### 2. Dependency Parser:

# STANFORD VERSION : More accurate but is too slow:
import os
os.environ["STANFORD_MODELS"] = "/home/meryem/Downloads/stanford-parser-full-2016-10-31"
os.environ["STANFORD_PARSER"] = "/home/meryem/Downloads/stanford-parser-full-2016-10-31"
from nltk.parse.stanford import StanfordDependencyParser
dep_parser=StanfordDependencyParser(model_path="/home/meryem/Downloads/stanford-parser-full-2016-10-31/edu/stanford/nlp/models/lexparser/englishPCFG.ser.gz")
dependency_trees = []
tweets_list = tweets_no_contractions['text'][0:10]
for i in range(0,len(tweets_list)):
    trees = [parse.tree() for parse in dep_parser.raw_parse(tweets_list[i])]
    result = dep_parser.raw_parse(tweets_list[i])
    dep = result.next()
    dependency_trees.append(list(dep.triples()))

On the sample data, we could not directly find interesting examples to show how our dependencies of interest (negation, adjectival complement and adverbial complement) are detected. That's why we will give a few examples that don't exist in the sample dataset but could be found in the whole dataset. 

dependency_trees_examples = []
trees_examples = []
examples = ['I am not happy','What a bad luck','I am struggling happily']
for i in range(0,len(examples)):
    trees_examples.append([parse.tree() for parse in dep_parser.raw_parse(examples[i])])
    result = dep_parser.raw_parse(examples[i])
    dep = result.next()
    dependency_trees_examples.append(list(dep.triples()))

dependency_trees_examples[0]

So here it has detected that happy depends on word not which cancels its emotion (happiness) as it has a negation dependency.  

dependency_trees_examples[1]

So here it has detected that luck depends on word bad which cancels its emotion (positive) as it has an adjectival modifier (amod) dependency.

dependency_trees_examples[2]

So here it has detected that struggling depends on word happily which cancels its emotion as it has an adjectival modifier (advmod) dependency. After that, struggling happily will have the emotion of depender "happily".

# TEMPORARY SOLUTION FOR DEPENDENCY PARSING:
nlp = spacy.load('en') # Loading nlp pipeline

# Creating docs
docs = []
# Joining text:
tweets_text = []
for i in range(0, len(tokenized_list)):
    space = u" "
    tweets_text.append(space.join(tokenized_list[i]))
tweets_text[0].encode("utf-8")
for i in range(0, len(tweets_text)):
    doc = nlp(tweets_text[i])
    docs.append(doc)

new_samples = []
for sample in docs:
    new_samples_sub = []
    for word in sample:
        new_samples_sub.append((unicode(word),word.pos_))
    new_samples.append(new_samples_sub)

### 3. Application of Syntactic Rules:

new_samples_syn,triple_dependencies_syn = apply_syntactic_rules(docs,new_samples)

i = 10
print "\n<<<< Original tweet text >>>\n"
print tweets_text[i]
print "\n<<<< Tweet after applying syntactic Rules >>>\n"
new_tweet = []
for (word,pos) in new_samples[i]:
    new_tweet.append(word)
print new_tweet

print tagged_tweets[0]
print "\n"
print new_samples[0]

## IV. Further Cleaning:

### 1. Named Entity Tagging:

import sys
reload(sys)
sys.setdefaultencoding("utf-8")
tweet_without_ne = remove_named_entities(new_samples)

### 2. Normalizing POS tag:

normalized_tags = normalize_pos_tags_words1(tweet_without_ne)
normalized_tags[1]

### 3. Removal of Punctuation and Stop words and Converting to Lower Case:

tagged_tweets_without = eliminate_stop_words_punct(normalized_tags)

### 4. Lemmatization:
To normalize terms to one common version

lemmatized_tweets = lemmatizer(tagged_tweets_without)

lemmatized_tweets_untag = lemmatizer_untagged(tagged_tweets_without)

lemmatized_tweets[1]

### 5.  Keeping only NAVA words

nava_tweets = keep_only_nava_words(lemmatized_tweets)
nava_tweets[0]

### 6. Lemmatizing Pre-cleaned Tokenized Tweets before any pre-processing:
We need this in order to train word2vec model as it will be impacted by the relative distance between words and we need the same lemmatized lower case version in order to calculate similarity scores for words that exist in nava tweets.

tokenized_lemmatized = lemmatizer_raw(normalize_pos_tags_words1(tagged_tweets))
tokenized_lemmatized[0]

### 6. Saving the lemmatized tokenized version and nava version of Tweets in a dataframe for later use:

nava_tweets_df = pd.DataFrame()
nava_tweets_df['Tokenized Lemmatized'] = tokenized_lemmatized
nava_tweets_df['Nava Representation'] = nava_tweets
nava_tweets_df.to_csv('../../Results/Sample Affective Representation.csv',index=False)

******************************* END *********************************************************
